requirements.txt
mss==10.1.0
requests==2.31.0
pytesseract==0.3.10
Pillow==10.2.0
joblib==1.3.2
scikit-learn==1.6.1